EtapeDto - modification des donnés en float

+++ Backoffice/rando/src/main/java/com/rando/dto/EtapeDto.java
@@ -10,9 +10,9 @@ public class EtapeDto {
 	@NotBlank(message = "Vous devez fournir une description !")
 	private String description;
 	@NotBlank(message = "Vous devez indiquer une latitude !")
-	private Integer latitude;
+	private Float latitude;
 	@NotBlank(message = "Vous devez indiquer une longitude !")
-	private Integer longitude;
+	private Float longitude;
 	@NotBlank(message = "Vous devez donner un nom à cette étape !")
 	private String nom;
 	private String qrCode;
@@ -45,25 +45,25 @@ public class EtapeDto {
 	/**
 	 * @return the latitude
 	 */
-	public Integer getLatitude() {
+	public Float getLatitude() {
 		return latitude;
 	}
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(Integer latitude) {
+	public void setLatitude(Float latitude) {
 		this.latitude = latitude;
 	}
 	/**
 	 * @return the longitude
 	 */
-	public Integer getLongitude() {
+	public Float getLongitude() {
 		return longitude;
 	}
 	/**
 	 * @param longitude the longitude to set
 	 */
-	public void setLongitude(Integer longitude) {
+	public void setLongitude(Float longitude) {
 		this.longitude = longitude;
 	}
 	/**
